backend/util: extract background handling code
